Added searchHelper to the MongoService so the other controllers can all use it without having to duplicate the code. Changed all instances of "searchHelper" to "MongoService.searchHelper" and deleted searchHelper from each file
